+++ lib/index.ts
@@ -1,1 +1,2 @@
 export { default } from './Chat';
+export { loadScript, renderChat } from './render-chat';