+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [0.1.1] - 2023-05-09
+
+- 🆕 `loadScript` and `renderChat` utilities to use Chat component in any JS project ([#10](https://github.com/enhancedocs/enhancedocs-chat/issues/10)).
+- 🛠 Vite: Build with ReactDOM as external/global and remove named export.
+
 ## [0.1.0] - 2023-05-08
 
 - 🆕 Answer feedback UX/UI ([#8](https://github.com/enhancedocs/enhancedocs-chat/issues/8)).